+++ be/src/olap/rowset/beta_rowset_writer.cpp
@@ -390,7 +390,7 @@ Status BetaRowsetWriter::_find_longest_consecutive_small_segment(
         SegCompactionCandidatesSharedPtr& segments) {
     segments = std::make_shared<SegCompactionCandidates>();
     // skip last (maybe active) segment
-    int32_t last_segment = _num_segment - 1;
+    int64_t last_segment = _num_segment - 1;
     size_t task_bytes = 0;
     uint32_t task_rows = 0;
     int32_t segid;
@@ -476,15 +476,15 @@ Status BetaRowsetWriter::_rename_compacted_segments(int64_t begin, int64_t end) 
     return Status::OK();
 }
 
-void BetaRowsetWriter::_clear_statistics_for_deleting_segments_unsafe(uint64_t begin,
-                                                                      uint64_t end) {
+void BetaRowsetWriter::_clear_statistics_for_deleting_segments_unsafe(uint32_t begin,
+                                                                      uint32_t end) {
     VLOG_DEBUG << "_segid_statistics_map clear record segid range from:" << begin << " to:" << end;
-    for (int i = begin; i <= end; ++i) {
+    for (uint32_t i = begin; i <= end; ++i) {
         _segid_statistics_map.erase(i);
     }
 }
 
-Status BetaRowsetWriter::_rename_compacted_segment_plain(uint64_t seg_id) {
+Status BetaRowsetWriter::_rename_compacted_segment_plain(uint32_t seg_id) {
     if (seg_id == _num_segcompacted) {
         ++_num_segcompacted;
         return Status::OK();
@@ -1044,7 +1044,7 @@ Status BaseBetaRowsetWriter::add_segment(uint32_t segment_id, const SegmentStati
         if (segment_id >= _segment_num_rows.size()) {
             _segment_num_rows.resize(segment_id + 1);
         }
-        _segment_num_rows[segid_offset] = segstat.row_num;
+        _segment_num_rows[segid_offset] = static_cast<uint32_t>(segstat.row_num);
     }
     VLOG_DEBUG << "_segid_statistics_map add new record. segment_id:" << segment_id
                << " row_num:" << segstat.row_num << " data_size:" << segstat.data_size
@@ -1053,7 +1053,7 @@ Status BaseBetaRowsetWriter::add_segment(uint32_t segment_id, const SegmentStati
     {
         std::lock_guard<std::mutex> lock(_segment_set_mutex);
         _segment_set.add(segid_offset);
-        while (_segment_set.contains(_num_segment)) {
+        while (_segment_set.contains(static_cast<uint32_t>(_num_segment))) {
             _num_segment++;
         }
     }
